requirements.txt
streamlit>=1.30.0
google-api-python-client>=2.120.0
google-auth>=2.20.0
python-dotenv>=1.0.0
requests>=2.31.0
scikit-learn>=1.3.0
pypdf>=4.2.0
python-docx>=1.1.0
python-pptx>=0.6.23
pandas>=2.2.0
openpyxl>=3.1.0
Pillow>=10.0.0
